Removed SSL Lookup
Instead of adding a new source, we can simply replace the old one with new API, so we can just replace the old source shodan with this and not adding shodandnsdb as a new one.
