+++ pkg/passive/sources.go
@@ -21,8 +21,7 @@ import (
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/passivetotal"
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/rapiddns"
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/securitytrails"
-	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/shodan/shodan"
-	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/shodan/shodandnsdb"
+	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/shodan"
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/sitedossier"
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/spyse"
 	"github.com/projectdiscovery/subfinder/pkg/subscraping/sources/sublist3r"
@@ -55,7 +54,6 @@ var DefaultSources = []string{
 	"rapiddns",
 	"securitytrails",
 	"shodan",
-	"shodandnsdb",
 	"sitedossier",
 	"spyse",
 	"sublist3r",
@@ -128,8 +126,6 @@ func (a *Agent) addSources(sources []string) {
 			a.sources[source] = &securitytrails.Source{}
 		case "shodan":
 			a.sources[source] = &shodan.Source{}
-		case "shodandnsdb":
-			a.sources[source] = &shodandnsdb.Source{}
 		case "sitedossier":
 			a.sources[source] = &sitedossier.Source{}
 		case "spyse":
